+++ app/lib/account/models.dart
@@ -345,6 +345,10 @@ class Consent extends db.Model {
   @db.StringProperty()
   String? fromUserId;
 
+  /// May be an `User.userId` or `support@pub.dev`.
+  @db.StringProperty()
+  String? fromAgent;
+
   @db.DateTimeProperty()
   DateTime? created;
 
@@ -364,6 +368,7 @@ class Consent extends db.Model {
 
   Consent.init({
     required this.fromUserId,
+    required this.fromAgent,
     required this.email,
     required this.kind,
     required this.args,
@@ -372,7 +377,7 @@ class Consent extends db.Model {
   }) {
     id = Ulid().toString();
     dedupId = consentDedupId(
-      fromUserId: fromUserId,
+      fromAgentId: fromAgent!,
       email: email,
       kind: kind,
       args: args!,
@@ -394,12 +399,12 @@ class Consent extends db.Model {
 
 /// Calculates the dedupId of a consent request.
 String consentDedupId({
-  required String? fromUserId,
+  required String fromAgentId,
   required String? email,
   required String? kind,
   required List<String> args,
 }) =>
-    [fromUserId, email, kind, ...args]
+    [fromAgentId, email, kind, ...args]
         .where((s) => s != null)
         .whereType<String>()
         .map(Uri.encodeComponent)