+++ CHANGELOG.md
@@ -5,6 +5,7 @@ AppEngine version, listed here to ease deployment and troubleshooting.
  * Bumped runtimeVersion to `2024.05.15`.
  * Upgraded runtime Dart SDK to `3.4.0`.
  * Upgraded dependencies.
+ * Note: started to populate `Consent.fromAgent` field.
 
 ## `20240514t110800-all`
  * Fixes copy icon issue.